backend/services/workflow-engine-service/requirements.txt
# Core FastAPI dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database dependencies
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.1

# Supabase dependencies
supabase==2.3.4
postgrest==0.13.2

# Google Cloud Healthcare API (commented out - install separately if needed)
# google-cloud-healthcare>=1.11.0
# google-auth>=2.25.2
# google-auth-oauthlib>=1.1.0
# google-auth-httplib2>=0.1.1

# GraphQL and Federation
graphql-core==3.2.3
strawberry-graphql[fastapi]==0.215.1
strawberry-graphql[debug-server]==0.215.1

# HTTP client
httpx==0.25.2
aiohttp==3.9.1

# Authentication and security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
cryptography==41.0.8  # PHI encryption for clinical workflows

# Workflow engine dependencies
pycamunda>=0.6.0  # Camunda REST API client # Camunda Cloud (Zeebe) client
requests>=2.25.0  # HTTP client for Camunda REST APIpyzeebe
pyzeebe
# Utilities
python-dotenv==1.0.0
structlog==23.2.0
rich==13.7.0

# Monitoring and observability
prometheus-client==0.17.1
opentelemetry-api==1.20.0
opentelemetry-sdk==1.20.0
opentelemetry-instrumentation-fastapi==0.41b0
opentelemetry-exporter-jaeger==1.20.0
opentelemetry-exporter-prometheus==1.12.0rc1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0
pytest-xdist==3.5.0
httpx==0.25.2
factory-boy==3.3.0

# Development
black==23.11.0
isort==5.12.0
flake8==6.1.0
